Remove  listener
